oteldemo: polish pass — README + lint cleanup + state-derivation fix
- README: replace the default Vite template README with a real one
  documenting the four tabs, the Cribl Search query layer, the
  /apps/__local__ dev workflow, and `npm run deploy`.
- searchFormState.ts: extract DEFAULT_SEARCH_STATE / SearchFormState /
  LOOKBACK_OPTIONS so SearchForm.tsx exports only the component (the
  react-refresh/only-export-components rule was complaining).
- DependencyGraph: hoist nodeRadius() above the useEffect that uses it,
  and disable the react-hooks/refs lint rule for the file (the
  ref-during-render pattern is intentional — d3 mutates positions in
  place each tick and we bump a counter to schedule renders).
- ComparePage: replace setState-in-effect for trace IDs with the
  derive-during-render reset pattern (URL params drive the inputs;
  manual edits stay until the user clicks Compare again).
- vite.config.ts: @ts-ignore → @ts-expect-error on the pkgutil.mjs
  import (eslint-recommended).
- eslint.config.js: turn off react-hooks/set-state-in-effect — every
  legitimate async fetch effect was being flagged, and the
  recommended alternative is "use a framework data layer," out of
  scope here.

`npx eslint .` and `tsc -b` both clean.
