+++ oteldemo/src/components/SearchForm.tsx
@@ -1,35 +1,7 @@
 import { useEffect, useState, type FormEvent } from 'react';
 import { listServices, listOperations } from '../api/search';
+import { LOOKBACK_OPTIONS, type SearchFormState } from './searchFormState';
 import s from './SearchForm.module.css';
-
-export interface SearchFormState {
-  service: string;
-  operation: string;
-  tags: string;
-  minDuration: string;
-  maxDuration: string;
-  limit: number;
-  lookback: string;
-}
-
-export const DEFAULT_SEARCH_STATE: SearchFormState = {
-  service: '',
-  operation: '',
-  tags: '',
-  minDuration: '',
-  maxDuration: '',
-  limit: 20,
-  lookback: '-1h',
-};
-
-const LOOKBACK_OPTIONS: Array<{ label: string; value: string }> = [
-  { label: 'Last 5 minutes', value: '-5m' },
-  { label: 'Last 15 minutes', value: '-15m' },
-  { label: 'Last 1 hour', value: '-1h' },
-  { label: 'Last 6 hours', value: '-6h' },
-  { label: 'Last 24 hours', value: '-24h' },
-  { label: 'Last 7 days', value: '-7d' },
-];
 
 interface Props {
   state: SearchFormState;
